dev: Improve UI unit test setup utilities
Relates to #3630 - other PRs can make use of this new utility, either by importing the file before any other imports or specifically importing `getWrapper` from the file and using the generated mock store and/or wrapper.

Change-type: patch
