+++ test/ui-setup.jsx
@@ -1,0 +1,64 @@
+/*
+ * Copyright (C) Balena.io - All Rights Reserved
+ * Unauthorized copying of this file, via any medium is strictly prohibited.
+ * Proprietary and confidential.
+ */
+
+import React from 'react'
+import {
+	Provider
+} from 'rendition'
+import {
+	Provider as ReduxProvider
+} from 'react-redux'
+import {
+	configure
+} from 'enzyme'
+import {
+	MemoryRouter
+} from 'react-router-dom'
+import configureStore from 'redux-mock-store'
+
+import Adapter from 'enzyme-adapter-react-16'
+
+configure({
+	adapter: new Adapter()
+})
+
+const browserEnv = require('browser-env')
+browserEnv([ 'window', 'document', 'navigator' ])
+
+const middlewares = []
+const mockStore = configureStore(middlewares)
+
+class XMLHttpRequest {}
+global.XMLHttpRequest = XMLHttpRequest
+
+class HowlerGlobal {}
+global.HowlerGlobal = HowlerGlobal
+
+class Howl {}
+global.Howl = Howl
+
+class Sound {}
+global.Sound = Sound
+
+export const getWrapper = (initialState = {}) => {
+	const store = mockStore(initialState)
+	return {
+		store,
+		wrapper: ({
+			children
+		}) => {
+			return (
+				<MemoryRouter>
+					<ReduxProvider store={store}>
+						<Provider>
+							{children}
+						</Provider>
+					</ReduxProvider>
+				</MemoryRouter>
+			)
+		}
+	}
+}
